+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `Behaviour Builder`:
     * All triggers have been changed to use two filters: filter for in subcontext data and filter for incoming data
     * New waiters for edited content
+    * New extension `BehaviourContext#followLocation`
 
 ## 0.35.8